backend/pyproject.toml
[project]
name = "mikir-kids-backend"
version = "0.1.0"
description = "FastAPI backend for Math Deliberate Practice MVP"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.115.0",
    "uvicorn[standard]==0.32.0",
    "python-dotenv==1.0.1",
    "pydantic==2.9.2",
    "pydantic-settings==2.5.2",
    "psycopg2-binary==2.9.10",
    "sqlalchemy==2.0.36",
    "alembic==1.13.2",
    "fsrs==0.1.0",
    "python-multipart==0.0.12",
]

[project.optional-dependencies]
test = [
    "requests==2.31.0",
]

[tool.hatch.build.targets.wheel]
packages = ["app"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

